refactor(inherited): rename & return better values
